refactor: move init code out of initFyo - initFyo just contains the global fyo object
