+++ src/setup/setupInstance.ts
@@ -11,8 +11,8 @@ import { AccountRootTypeEnum } from 'models/baseModels/Account/types';
 import { AccountingSettings } from 'models/baseModels/AccountingSettings/AccountingSettings';
 import { numberSeriesDefaultsMap } from 'models/baseModels/Defaults/Defaults';
 import { ModelNameEnum } from 'models/types';
-import { initializeInstance, setCurrencySymbols } from 'src/initFyo';
 import { createRegionalRecords } from 'src/regional';
+import { initializeInstance, setCurrencySymbols } from 'src/utils/initialization';
 import { getRandomString } from 'utils';
 import { defaultUOMs } from 'utils/defaults';
 import { getCountryCodeFromCountry, getCountryInfo } from 'utils/misc';
